Text display mostly functional
The main change is that now I have the window text
displaying properly. I am also able to grab an
arbitrary font; it'll be Courier New for now, mostly
for the monospace. Next thing will be to make the
string it displays be created by the
AsciiDisplaySystem.

The other thing I've done is make the window not
resizable. This is just so I don't have to resize
the text or other nonsense.

Signed-off-by: CKraniak <ckraniak@live.com>