Dispatcher completed, beginning CesIo
The Dispatcher class is fully implemented. It will
need a comprehensive test function, since it's a
complicated class with a lot of potential for bugs
and a lot of potential to become VERY core code.

In particular, the way event handlers are declared
may be changed, because the macro way is probably
not going to work for any real needs.

I also spruced up the dispatcher with quite a few
comments. It should help quite a bit. License
statements as well. I need to get GPLv3 and put it
in the root directory.

The next step is to fix the CES system, which means:
 - being able to read in from files
 - creating entity types (not actual entities)
 --- the files are like entity templates
 - populating entities with components
 - having systems match the entities they need
 - SystemGroup object
 - systems communicating via events, not direct handles

We can then clean up the MLInterface a bit.

After this, the priority switches to making a stub
world builder that just loads static maps. Then,
collision detection against terrain. And so it goes.

Signed-off-by: CKraniak <ckraniak@live.com>