Codechange: Move MLInterface into its own directory
A couple of less important changes in the ASCII rendering
code, mostly involving structure. I am switching to
a "driver" structure in the graphics department, where
said driver provides the platform-specific backend
a renderer needs to function.

Main change was getting MLInterface out of main, and
completely restructuring both main and MLInterface.
The Windows code is now in a subclass of MLInterface,
so extending to different platfors should be a bit
more straightforward now.