Add: Expand SystemGroup, CESystem, create LoadEntityCESystem
New design direction (or rather, have decided how to handle having a master
list). Instead of a SystemGroup or main holding an authoritative list of
entities and components, a special system will do it. The special system
gets to be responsible for both loading and cacheing everything it loads.

Actaully ... no. I need a special MasterListCESystem or somesuch. Something
specifically made with that function in mind, not loading AND controlling.

We'll see how it goes.