Beginning connection and debug of dispatcher in CES
Now that the dispatcher is usable, it needs to be merged into the currently
functioning code that should have been using an event system anyway.

Changes have been made to the CESystem class to facilitate an
onDispatcherAvailable type of "event", so systems can connect their handlers
at the appropriate time.

Some spart pointer usage was fixed to prevent double deletes on program
exit.

Also, a "this" pointer was added to handlers to facilitate handler-
specific data.