Finalization of CES design, moved platform-dependent code to containment
- CES design given missing link: ECList
- Directory class fleshed out, working so far
---  Note: required for CesIo, which is required for the CES as a whole
- Windows code mostly moved to /platform/windows now
--- Affects: utility.h, Directory
--- Remaining: GLFunctionFinder (to do WAY later)
- Some work to integrate events into reacting to input
--- GameGrid correctly tracking player, but not applying motions (still
    done manually by WinProc (?)). Has been waiting on better
    implementation of Component code so the queries are less clunky. This
    is why the ECList was so important: it's needed to finish this part
    (and everything else) off.

Once I finish the CES code, I can finish the collision detector and wean
the code off of manual positioning from WinProc and into event-based
movement. To finish the CES code, the ECList design element had been needed
and is now available, and also I need to be able to load the stuff from
files and test it out. So I need CesIo, which needed Directory, which is
now made.

The platform code consolidation has been something that's been coming
for a while anyway, so I went and just did it.

So in order:
- Finish CesIo and test
- Finish EntityType / EntityFactory and test
- Finish ECList / EntityList / ComponentList and test
- Finish Collision System (at least a stub) and test
- Unhook WinProc from CES completely and have the systems move themselves
- Finalize CES / clean all of it up

Then we will upversion to prealpha4-4.

Signed-off-by: CKraniak <ckraniak@live.com>